third_party/WebKit/LayoutTests/inspector/tracing/decode-resize-expected.txt

Tests the instrumentation of a DecodeImage and ResizeImage events


event: Decode Image
imageURL: .../inspector/tracing/resources/test.bmp
backendNodeId: present
event: Decode Image
imageURL: .../inspector/tracing/resources/test.gif
backendNodeId: present
event: Decode Image
imageURL: .../inspector/tracing/resources/test.ico
backendNodeId: present
event: Decode Image
imageURL: .../inspector/tracing/resources/test.jpg
backendNodeId: present
event: Decode Image
imageURL: .../inspector/tracing/resources/test.png
backendNodeId: present
event: Decode Image
imageURL: .../inspector/tracing/resources/test.webp
backendNodeId: present
event: Decode Image
imageURL: .../inspector/tracing/resources/big.png
backendNodeId: present
